Add 93% test coverage for NavBar component - Rename test file to index
